requirements.txt
streamlit==1.24.1
langchain
python-dotenv
tiktoken
langchain_community
groq
faiss-cpu
sentence-transformers
PyPDF2
pymupdf
langchain-groq
numpy<2
pandas
pybind11>=2.12
